Move create_dataframe to construction
